+++ src/modules/system/preload-templates.js
@@ -98,6 +98,7 @@ export default function preloadHandlebarsTemplates() {
     `systems/${game.system.id}/templates/chat/cpr-rollcard-modifiers.hbs`,
 
     // Dialog: Partials
+    `systems/${game.system.id}/templates/dialog/cpr-cyberpunkred-com-import.hbs`,
     `systems/${game.system.id}/templates/dialog/cpr-default-prompt.hbs`,
     `systems/${game.system.id}/templates/dialog/cpr-damage-application-prompt.hbs`,
     `systems/${game.system.id}/templates/dialog/cpr-dialog-buttons.hbs`,